requirements.txt
fastapi
uvicorn
dotenv
snowflake-connector-python
requests
python-geohash
datetime